id gets removed on resourcerelation calls
In the Resource Relation handler, the id payload gets used and set into a foreign key relation. The old original key however, was not removed and was kept in the payload. this ended up in unneccesary data being sent to the endpoints and the endpoints having to cope with the id even when it was not in its contract.